+++ SEDAPExpress/src/main/java/de/bundeswehr/mese/sedapexpress/messages/COMMAND.java
@@ -194,7 +194,7 @@ public class COMMAND extends SEDAPExpressMessage {
      * @param cmdType
      * @param cmdTypeDependentParameters
      */
-    public COMMAND(Short number, Long time, String sender, Classification classification, Acknowledgement acknowledgement, String mac, String recipient, Short cmdId, CommandFlag cmdFlag, CommandType cmdType,
+    public COMMAND(Byte number, Long time, String sender, Classification classification, Acknowledgement acknowledgement, String mac, String recipient, Short cmdId, CommandFlag cmdFlag, CommandType cmdType,
 	    List<String> cmdTypeDependentParameters) {
 
 	super(number, time, sender, classification, acknowledgement, mac);
