+++ SEDAPExpress/src/test/java/de/bundeswehr/mese/sedapexpress/messages/OWNUNITTest.java
@@ -44,9 +44,9 @@ class OWNUNITTest {
     @Test
     final void testConstructorValues() {
 
-	final OWNUNIT ownunit = new OWNUNIT((short) 11, 456465543L, "22AA", Classification.Unclas, Acknowledgement.TRUE, "4389F10D", 77.88d, -10.12d, 5577.0d, 33.44d, 55.66d, 1.1d, -2.2d, 3.3d, "Ownunit", "SFGPIB----H----".toCharArray());
+	final OWNUNIT ownunit = new OWNUNIT((byte) 11, 456465543L, "22AA", Classification.Unclas, Acknowledgement.TRUE, "4389F10D", 77.88d, -10.12d, 5577.0d, 33.44d, 55.66d, 1.1d, -2.2d, 3.3d, "Ownunit", "SFGPIB----H----".toCharArray());
 
-	Assertions.assertEquals((short) 11, ownunit.getNumber());
+	Assertions.assertEquals((byte) 11, ownunit.getNumber());
 	Assertions.assertEquals(456465543L, ownunit.getTime());
 	Assertions.assertEquals("22AA", ownunit.getSender());
 	Assertions.assertEquals(Classification.Unclas, ownunit.getClassification());
@@ -72,7 +72,7 @@ class OWNUNITTest {
 
 	OWNUNIT ownunit = new OWNUNIT(message);
 
-	Assertions.assertEquals((short) 0x11, ownunit.getNumber());
+	Assertions.assertEquals((byte) 0x11, ownunit.getNumber());
 	Assertions.assertEquals(0x1B351C87L, ownunit.getTime());
 	Assertions.assertEquals("22AA", ownunit.getSender());
 	Assertions.assertEquals(Classification.Unclas, ownunit.getClassification());
@@ -93,7 +93,7 @@ class OWNUNITTest {
 
 	ownunit = new OWNUNIT(message);
 
-	Assertions.assertEquals((short) 0x5E, ownunit.getNumber());
+	Assertions.assertEquals((byte) 0x5E, ownunit.getNumber());
 	Assertions.assertEquals(0x661D4410L, ownunit.getTime());
 	Assertions.assertEquals("66A3", ownunit.getSender());
 	Assertions.assertEquals(Classification.Restricted, ownunit.getClassification());
@@ -117,7 +117,7 @@ class OWNUNITTest {
 
 	OWNUNIT ownunit = new OWNUNIT(it);
 
-	Assertions.assertEquals((short) 0x11, ownunit.getNumber());
+	Assertions.assertEquals((byte) 0x11, ownunit.getNumber());
 	Assertions.assertEquals(0x1B351C87L, ownunit.getTime());
 	Assertions.assertEquals("22AA", ownunit.getSender());
 	Assertions.assertEquals(Classification.Unclas, ownunit.getClassification());
